refactor(neovim): restructure plugins to use per file configs
