--- neovim/.config/nvim/lua/config/plugin-settings/colorizer.lua
@@ -1,3 +1,0 @@
-local success, colorizer = pcall(require, 'colorizer')
-if not success then return end
-colorizer.setup()
