--- neovim/.config/nvim/lua/config/plugin-settings/hologram.lua
@@ -1,3 +1,0 @@
-local success, hologram = pcall(require, 'hologram')
-if not success then return end
-hologram.setup({ auto_display = true })
