--- neovim/.config/nvim/lua/config/plugin-settings/indent-blankline.lua
@@ -1,12 +1,0 @@
-local loaded, indent_blankline = pcall(require, 'ibl')
-if not loaded then return end
-
-vim.opt.list = true
-vim.opt.termguicolors = true
-
-indent_blankline.setup({
-  debounce = 100,
-  -- Special Characters: '│' '█' '▏'
-  indent = { char = "▏", highlight = { "Whitespace", "NonText" }, },
-  whitespace = { highlight = { "Whitespace", "NonText" } },
-})
