--- neovim/.config/nvim/lua/config/plugin-settings/nvim-cmp.lua
@@ -1,165 +1,0 @@
-vim.o.completeopt = "menu,menuone,noselect"
-
--- Setup nvim-cmp and luasnip
-local cmp_status_ok, cmp = pcall(require, "cmp")
-if not cmp_status_ok then return end
-local snip_status_ok, luasnip = pcall(require, "luasnip")
-if not snip_status_ok then return end
-
-local kind_icons = {
-  Text          = "",
-  Method        = "m",
-  Function      = "",
-  Constructor   = "",
-  Field         = "",
-  Variable      = "",
-  Class         = "",
-  Interface     = "",
-  Module        = "",
-  Property      = "",
-  Unit          = "",
-  Value         = "",
-  Enum          = "",
-  Keyword       = "",
-  Snippet       = "",
-  Color         = "",
-  File          = "",
-  Reference     = "",
-  Folder        = "",
-  EnumMember    = "",
-  Constant      = "",
-  Struct        = "",
-  Event         = "",
-  Operator      = "",
-  TypeParameter = "",
-}
-
-local check_backspace = function()
-  local col = vim.fn.col "." - 1
-  return col == 0 or vim.fn.getline("."):sub(col, col):match "%s"
-end
-
-cmp.setup({
-  snippet = {
-    expand = function(args)
-      luasnip.lsp_expand(args.body) -- For `luasnip` users.
-    end,
-  },
-  mapping = {
-    ["<C-k>"] = cmp.mapping.select_prev_item(),
-    ["<C-j>"] = cmp.mapping.select_next_item(),
-    ["<C-b>"] = cmp.mapping(cmp.mapping.scroll_docs(-1), { "i", "c" }),
-    ["<C-f>"] = cmp.mapping(cmp.mapping.scroll_docs(1), { "i", "c" }),
-    ["<C-Space>"] = cmp.mapping(cmp.mapping.complete(), { "i", "c" }),
-    ["<C-y>"] = cmp.config.disable, -- Specify `cmp.config.disable` if you want to remove the default `<C-y>` mapping.
-    ["<C-e>"] = cmp.mapping {
-      i = cmp.mapping.abort(),
-      c = cmp.mapping.close(),
-    },
-    -- Accept currently selected item. If none selected, `select` first item.
-    -- Set `select` to `false` to only confirm explicitly selected items.
-    ["<CR>"] = cmp.mapping.confirm { select = true },
-    ["<Tab>"] = cmp.mapping(function(fallback)
-      if cmp.visible() then
-        cmp.select_next_item()
-      elseif luasnip.expandable() then
-        luasnip.expand()
-      elseif luasnip.expand_or_jumpable() then
-        luasnip.expand_or_jump()
-      elseif check_backspace() then
-        fallback()
-      else
-        fallback()
-      end
-    end, {
-        "i",
-        "s",
-      }),
-    ["<S-Tab>"] = cmp.mapping(function(fallback)
-      if cmp.visible() then
-        cmp.select_prev_item()
-      elseif luasnip.jumpable(-1) then
-        luasnip.jump(-1)
-      else
-        fallback()
-      end
-    end, {
-      "i",
-      "s",
-    }),
-  },
-  formatting = {
-    fields = { "kind", "abbr", "menu" },
-    format = function(entry, vim_item)
-      -- Kind icons
-      vim_item.kind = string.format("%s ", kind_icons[vim_item.kind])
-      vim_item.menu = ({
-        nvim_lsp = "[LSP]",
-        luasnip = "[Snippet]",
-        buffer = "[Buffer]",
-        path = "[Path]",
-      })[entry.source.name]
-      return vim_item
-    end,
-  },
-  sources = {
-    { name = "nvim_lsp" },
-    { name = "luasnip" },
-    { name = "buffer" },
-    { name = "path" },
-    { name = 'nvim_lsp_signature_help' },
-  },
-  confirm_opts = {
-    behavior = cmp.ConfirmBehavior.Replace,
-    select = false,
-  },
-  window = {
-    documentation = {
-      -- border = { "╭", "─", "╮", "│", "╯", "─", "╰", "│" },
-      border = { " ", " ", " ", " ", " ", " ", " ", " " },
-    },
-  },
-  experimental = {
-    ghost_text = true,
-    native_menu = false,
-  },
-})
-
--- Set configuration for specific filetype.
-cmp.setup.filetype('gitcommit', {
-  sources = cmp.config.sources(
-    {
-      { name = 'cmp_git' }, -- You can specify the `cmp_git` source if you installed it.
-    },
-    {
-      { name = 'buffer' },
-    }
-  )
-})
-
--- Use buffer source for `/` (if you enabled `native_menu`, this won't work anymore).
-cmp.setup.cmdline('/', {
-  sources = {
-    { name = 'buffer' }
-  }
-})
-
--- Use cmdline & path source for ':' (if you enabled `native_menu`, this won't work anymore).
-cmp.setup.cmdline(':', {
-  sources = cmp.config.sources(
-    {
-      { name = 'path' }
-    },
-    {
-      { name = 'cmdline' }
-    }
-  )
-})
-
--- Setup lspconfig.
-local capabilities = require('cmp_nvim_lsp').default_capabilities(vim.lsp.protocol.make_client_capabilities())
-local lspconfig_loaded, lspconfig = pcall(require, 'lspconfig')
-if not lspconfig_loaded then return end
-lspconfig['jsonls'].setup {
-  capabilities = capabilities
-}
