--- neovim/.config/nvim/lua/config/plugin-settings/nvim-highlight-colors.lua
@@ -1,6 +1,0 @@
-local success, highlight_colors = pcall(require, "nvim-highlight-colors")
-if not success then return end
-
-highlight_colors.setup{
-  render = "background" -- "background", "first_column" or "foreground"
-}
