--- neovim/.config/nvim/lua/config/plugin-settings/nvim-navic.lua
@@ -1,37 +1,0 @@
-local success, navic = pcall(require, 'nvim-navic')
-if not success then return end
-
-navic.setup {
-  icons = {
-    Array         = " ",
-    Boolean       = ' ',
-    Class         = ' ',
-    Constant      = " ",
-    Constructor   = " ",
-    Enum          = ' ',
-    EnumMember    = ' ',
-    Event         = " ",
-    Field         = " ",
-    File          = " ",
-    Function      = " ",
-    Interface     = ' ',
-    Key           = " ",
-    Method        = " ",
-    Module        = " ",
-    Namespace     = " ",
-    Null          = "ﳠ ",
-    Number        = " ",
-    Object        = " ",
-    Operator      = ' ',
-    Package       = ' ',
-    Property      = " ",
-    String        = " ",
-    Struct        = ' ',
-    TypeParameter = ' ',
-    Variable      = ' ',
-  },
-  highlight = true,
-  separator = " > ",
-  depth_limit = 0,
-  depth_limit_indicator = "..",
-}
