--- neovim/.config/nvim/lua/config/plugin-settings/symbols-outline.lua
@@ -1,3 +1,0 @@
-local success, symbols_outline = pcall(require, 'symbols-outline')
-if not success then return end
-symbols_outline.setup()
