--- neovim/.config/nvim/lua/config/plugin-settings/telescope-media-files.lua
@@ -1,3 +1,0 @@
-local success, telescope = pcall(require, 'telescope')
-if not success then return end
-telescope.load_extension('media_files')
