+++ neovim/.config/nvim/lua/config/plugins/extra.lua
@@ -1,0 +1,15 @@
+return {
+  {
+    'nvim-lua/popup.nvim',
+    lazy = true
+  },
+  {
+    'jkramer/vim-checkbox',
+    keys = '<leader>tt'
+  },
+  {
+    'godlygeek/tabular',
+    keys = '<leader>a',
+    cmd = 'Tabularize'
+  },
+}
