+++ neovim/.config/nvim/lua/config/utils/border.lua
@@ -19,4 +19,25 @@ else
   }
 end
 
-return borders
+return {
+  borderchars = {
+    borders[2],
+    borders[6],
+    borders[8],
+    borders[4],
+    borders[1],
+    borders[3],
+    borders[9],
+    borders[7],
+  },
+  borders = {
+    { borders[1], "FloatBorder" },
+    { borders[2], "FloatBorder" },
+    { borders[3], "FloatBorder" },
+    { borders[6], "FloatBorder" },
+    { borders[9], "FloatBorder" },
+    { borders[8], "FloatBorder" },
+    { borders[7], "FloatBorder" },
+    { borders[4], "FloatBorder" },
+  },
+}
